requirements.txt
celery==4.4.7
python-dotenv==0.14.0
redis==3.5.3
